refactor app module completed
